Time: 64 ms (5.07%) | Memory: 95.9 MB (24.59%) - LeetSync

+++ 4-median-of-two-sorted-arrays/median-of-two-sorted-arrays.cpp
@@ -27,7 +27,7 @@ public:
         }
         int n=v.size();
         if(n%2==0){
-            return (double)(v[n/2]+v[(n/2)-1])/2;
+            return (float)(v[n/2]+v[(n/2)-1])/2;
         }
         else{
             return v[n/2];
